% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclogit-rebase.R
\name{rebase}
\alias{rebase}
\title{Change baseline category of multinomial logit or similar model}
\usage{
rebase(object, to, ...)
}
\arguments{
\item{object}{a statistical model object}

\item{to}{usually, a string; the baseline category}
}
\description{
`rebase` returns an model object that is equivalent to the one
given as argument but differs in parameterization
}
