\name{Transport}
\alias{Transport}
\title{Choice of Means of Transport}
\description{This is an artificial data set on
choice of means of transport based on cost and walking distance.
}
\usage{data(Transport)}
\format{A data frame containing the following variables:

\describe{
	\item{transport}{means of transportation that can be chosen.}
	\item{suburb}{identifying number for each suburb}
	\item{distance}{walking distance to bus or train station}
	\item{cost}{cost of each means of transportation}
	\item{working}{size of working population of each suburb}
	\item{prop.true}{true choice probabilities}
	\item{resp}{choice frequencies of means of transportation}
}
}

\keyword{datasets}
